refactor: move and rename Menu and MenuBackdrop
