+++ src/components/AppComponents/AppHeaderMenu/index.ts
@@ -1,0 +1,3 @@
+export { default as AppHeaderMenu } from "./AppHeaderMenu";
+export type { TransitionProps } from "./AppHeaderMenu";
+export { SideMenu } from "./AppHeaderMenu";
