+++ src/components/AppComponents/MenuBackdrop/MenuBackdrop.test.tsx
@@ -1,9 +1,10 @@
 import mockRouter from "next-router-mock";
 
-import { MenuProvider } from "../../context/Menu";
-import renderWithTheme from "../../__tests__/__helpers__/renderWithTheme";
-
 import MenuBackdrop from "./MenuBackdrop";
+
+import { MenuProvider } from "@/context/Menu";
+import renderWithTheme from "@/__tests__/__helpers__/renderWithTheme";
+
 
 jest.mock("next/dist/client/router", () => require("next-router-mock"));
