--- src/components/Menu/index.ts
@@ -1,1 +1,0 @@
-export { default as Menu } from "./Menu";
